test: split apart the search tests
